Moving more stuff to CATCH
