+++ session06/cpp/collective.cc
@@ -1,43 +1,44 @@
 #include <mpi.h>
-#include <iostream>
+#define CATCH_CONFIG_RUNNER
+#include "catch.hpp"
 
-/// "functions"
-void root_broadcasts(std::string const message) {
-    int const error = MPI_Bcast(
-            (void*) message.c_str(), message.size(), MPI_CHAR, 0,
-            MPI_COMM_WORLD
-    );
-    if(error != MPI_SUCCESS) throw;
-}
-
-std::string others_get_it() {
-    char buffer[256];
-    int const error = MPI_Bcast(buffer, 256, MPI_CHAR, 0, MPI_COMM_WORLD);
-    if(error != MPI_SUCCESS) throw;
-    return buffer;
-}
-
-/// "dummy"
-int main(int argc, char * argv[]) {
-    MPI_Init (&argc, &argv);
+TEST_CASE("Collective communications") {
 
     int rank, size;
     MPI_Comm_rank (MPI_COMM_WORLD, &rank);
     MPI_Comm_size (MPI_COMM_WORLD, &size);
 
-    /// "main"
-    std::string message = "";
-    if(rank == 0) root_broadcasts("Listen up folks.");
-    else message = others_get_it();
+    SECTION("Collective broadcast") {
+      /// "broadcast"
+      std::string const peace = "I come in peace!";
+      std::string message = "";
+      int error;
+      if(rank == 0) {
+          message = peace;
+          error = MPI_Bcast(
+             (void*) peace.c_str(), peace.size() + 1, MPI_CHAR, 0, MPI_COMM_WORLD);
+      } else {
+          char buffer[256];
+          int const error = MPI_Bcast(buffer, 256, MPI_CHAR, 0, MPI_COMM_WORLD);
+          message = std::string(buffer);
+      }
 
-    for(int i(1); i < size; ++i) {
-        if(rank == i)
-            std::cout << "I am " << rank
-                << ", and I approve this message: " << message << "\n";
-        MPI_Barrier(MPI_COMM_WORLD);
+      /// "tests"
+      for(int i(0); i < size; ++i) {
+          if(rank == i) {
+            INFO("Current rank is " << rank);
+            REQUIRE(error == MPI_SUCCESS);
+            CHECK(message == peace);
+          }
+          MPI_Barrier(MPI_COMM_WORLD);
+      }
+      /// "dummy"
     }
+}
 
-    /// "dummy"
+int main(int argc, char * argv[]) {
+    MPI_Init (&argc, &argv);
+    int result = Catch::Session().run(argc, argv);
     MPI_Finalize();
-    return 0;
+    return result;
 }
